+++ packages/editor/src/parser/html/html.ts
@@ -1,0 +1,6 @@
+import unified from 'unified';
+import parse from 'rehype-dom-parse';
+
+export const processor = unified()
+  .use(parse)
+  .data('settings', { fragment: true });